Added minutes to angle timers
- Diagnose with re<S><DD><MM> S=sign, D=degree, M=minute
- Value range reduced to 63°59'

+++ src/Timer.cpp
@@ -237,7 +237,7 @@ sTime *Timer::getSunInfo(uint8_t iSunInfo)
         return NULL;
 }
 
-void Timer::getSunDegree(uint8_t iSunInfo, int8_t iDegree, sTime *eSun)
+void Timer::getSunDegree(uint8_t iSunInfo, double iDegree, sTime *eSun)
 {
     double rise, set;
     // sunrise/sunset calculation
